pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "joblib>=1.4.2",
    "openai>=1.75.0",
    "pandas>=2.2.3",
    "pdfplumber>=0.11.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "python-docx>=1.1.2",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "spacy>=3.8.5",
    "streamlit-extras>=0.7.1",
    "streamlit>=1.44.1",
    "trafilatura>=2.0.0",
    "twilio>=9.5.2",
    "uvicorn>=0.34.2",
]
